Правки по замечаниям Спринт 7. Итерация 1. Создал базовый класс для юнит-тестов менеджеров. Разрешил менять статус Эпика напрямую для целей точного чтения загружаемого файла (все равно изменить статус эпика извне не получится из-за клонирования) Сделал в TaskUtils приватный конструктор Реализовал загрузку задач из файла с прямой записью в структуру, без использования методов создания. Добавил заглушечный тип и статус задачи, чтобы не отлавливать исключение в парсинге
